Move integration tests to the correct namespace
