+++ tests/Integration/Database/SqlServer/DatabaseEloquentSqlServerIntegrationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Illuminate\Tests\Database;
+namespace Illuminate\Tests\Integration\Database\SqlServer;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Schema\Blueprint;
